+++ canaille/app/configuration.py
@@ -3,6 +3,7 @@ import re
 import smtplib
 import socket
 import textwrap
+from dataclasses import dataclass
 
 from pydantic import BaseModel as PydanticBaseModel
 from pydantic import ValidationError
@@ -166,6 +167,12 @@ class ConfigurationException(Exception):
     pass
 
 
+@dataclass
+class CheckResult:
+    message: str
+    success: bool | None = None
+
+
 def setup_config(app, config=None, env_file=None, env_prefix=""):
     from canaille.oidc.installation import install
 
@@ -206,16 +213,25 @@ def setup_config(app, config=None, env_file=None, env_prefix=""):
 
 
 def check_network_config(config):
+    """Perform various network connection to services described in the configuration file."""
     from canaille.backends import Backend
 
-    Backend.instance.check_network_config(config)
+    results = [Backend.instance.check_network_config(config)]
+
     if smtp_config := config["CANAILLE"]["SMTP"]:
-        test_smtp_configuration(smtp_config)
+        results.append(check_smtp_connection(smtp_config))
+    else:
+        results.append(CheckResult(message="No SMTP server configured"))
+
     if smpp_config := config["CANAILLE"]["SMPP"]:
-        test_smpp_configuration(smpp_config)
+        results.append(check_smpp_connection(smpp_config))
+    else:
+        results.append(CheckResult(message="No SMPP server configured"))
+
+    return results
 
 
-def test_smtp_configuration(config):
+def check_smtp_connection(config) -> str:
     host = config["HOST"]
     port = config["PORT"]
     try:
@@ -228,21 +244,31 @@ def test_smtp_configuration(config):
                     user=config["LOGIN"],
                     password=config["PASSWORD"],
                 )
-    except (socket.gaierror, ConnectionRefusedError) as exc:
-        raise ConfigurationException(
-            f"Could not connect to the SMTP server '{host}' on port '{port}'"
-        ) from exc
+    except (socket.gaierror, ConnectionRefusedError):
+        return CheckResult(
+            message=f"Could not connect to the SMTP server '{host}' on port '{port}'",
+            success=False,
+        )
 
-    except smtplib.SMTPAuthenticationError as exc:
-        raise ConfigurationException(
-            f"SMTP authentication failed with user '{config['LOGIN']}'"
-        ) from exc
+    except smtplib.SMTPAuthenticationError:
+        return CheckResult(
+            message=f"SMTP authentication failed with user '{config['LOGIN']}'",
+            success=False,
+        )
 
     except smtplib.SMTPNotSupportedError as exc:
-        raise ConfigurationException(exc) from exc
+        return CheckResult(
+            message=str(exc),
+            success=False,
+        )
+
+    return CheckResult(
+        message="Successful SMTP connection",
+        success=True,
+    )
 
 
-def test_smpp_configuration(config):
+def check_smpp_connection(config):
     import smpplib
 
     host = config["HOST"]
@@ -254,12 +280,21 @@ def test_smpp_configuration(config):
                 client.bind_transmitter(
                     system_id=config["LOGIN"], password=config["PASSWORD"]
                 )
-    except smpplib.exceptions.ConnectionError as exc:
-        raise ConfigurationException(
-            f"Could not connect to the SMPP server '{host}' on port '{port}'"
-        ) from exc
+    except smpplib.exceptions.ConnectionError:
+        return CheckResult(
+            success=False,
+            message=f"Could not connect to the SMPP server '{host}' on port '{port}'",
+        )
     except smpplib.exceptions.UnknownCommandError as exc:  # pragma: no cover
-        raise ConfigurationException(exc) from exc
+        return CheckResult(
+            message=str(exc),
+            success=False,
+        )
+
+    return CheckResult(
+        message="Successful SMPP connection",
+        success=True,
+    )
 
 
 def sanitize_rst_text(text: str) -> str: